add lab 03-2, fix guess number

+++ week3/GotFlip.php
@@ -6,7 +6,7 @@
 
 <body>
     <?php
-    $rand = ((double) microtime() * 1000000);
+    $rand = ((float) microtime() * 1000000);
     $flip = rand(0, 1);
     $pick = $_POST["pick"];
     if ($flip == 0 && $pick == 0) {
